Consolidate card form test helpers
Committed-By-Agent: codex
